+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Initial connector which outputs all analytics into a single elasticsearch index
 - Enabled the ES connecter to save all output to a rolling index rather than a static index
+- Add support to the Prometheus pump to exclude metric families from exposition 
 
 ## v0.1